requirements.txt
asgiref==3.8.1
cffi<2.0.0,>=1.16.0
cryptography>=42.0.0
distro==1.9.0
dotenv==0.9.9
gunicorn==21.2.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
openpyxl==2.6.2
six==1.17.0
sqlparse==0.5.3
tabula-py==2.10.0
tzdata==2025.1
Django==5.1.4
mysqlclient==2.2.4
PyJWT==2.8.0
google-auth==2.40.1
requests==2.31.0
google-api-python-client==2.167.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1